% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio01.R
\name{bio01}
\alias{bio01}
\title{BIO01 Mean Annual Temperature}
\usage{
bio01(tas)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO01 Mean Annual Temperature
}
