% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio02.R
\name{bio02}
\alias{bio02}
\title{BIO02 Mean Annual Diurnal Range}
\usage{
bio02(tmax, tmin)
}
\arguments{
\item{tmax}{raster stack of monthly maximum temperatures.}

\item{tmin}{raster stack of monthly minimum temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO02 Mean Annual Diurnal Range
}
