% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio03.R
\name{bio03}
\alias{bio03}
\title{BIO03 Mean Annual Isothermality}
\usage{
bio03(tmax, tmin)
}
\arguments{
\item{tmax}{raster stack of monthly maximum temperatures.}

\item{tmin}{raster stack of monthly minimum temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO03 Mean Annual Isothermality
}
\details{
In contrast to CHELSA, this is not multiplied by 100
}
