% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio04.R
\name{bio04}
\alias{bio04}
\title{BIO04 Annual Mean Temperature Seasonality}
\usage{
bio04(tas)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO04 Annual Mean Temperature Seasonality
}
\details{
Standard deviation of the monthly mean temperatures.
In contrast to CHELSA, this is not multiplied by 100
}
