% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio05.R
\name{bio05}
\alias{bio05}
\title{BIO05 Maximum Temperature of Warmest Month}
\usage{
bio05(tmax)
}
\arguments{
\item{tmax}{raster stack of monthly average temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO05 Maximum Temperature of Warmest Month
}
