% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio06.R
\name{bio06}
\alias{bio06}
\title{BIO06 Minimum Temperature Coldest Month}
\usage{
bio06(tmin)
}
\arguments{
\item{tmin}{raster stack of monthly average temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO06 Minimum Temperature Coldest Month
}
