% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio07.R
\name{bio07}
\alias{bio07}
\title{BIO07 Mean Annual Temperature range}
\usage{
bio07(tmax, tmin)
}
\arguments{
\item{tmax}{raster stack of monthly maximum temperatures.}

\item{tmin}{raster stack of monthly minimum temperatures.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO07 Mean Annual Temperature range
}
