% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio08.R
\name{bio08}
\alias{bio08}
\alias{cpp_bio08}
\title{Average Value of the Wettest Quarter}
\usage{
cpp_bio08(tas, pr)

bio08(tas, pr, also.quarter = FALSE)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}

\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average temperature of the highest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Average Value of the Wettest Quarter

BIO08 Annual Mean Temperature of Wettest Quarter
}
