% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio09.R
\name{bio09}
\alias{bio09}
\alias{cpp_bio09}
\title{Mean Temperature of Driest Quarter}
\usage{
cpp_bio09(tas, pr)

bio09(tas, pr, also.quarter = FALSE)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}

\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average temperature of the driest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Mean Temperature of Driest Quarter

BIO09 Mean Temperature of Driest Quarter
}
