% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio-function.R
\name{bio10}
\alias{bio10}
\title{BIO10 Mean Temperature of the Warmest Quarter}
\usage{
bio10(s, also.quarter = FALSE)
}
\arguments{
\item{s}{raster stack of monthly temperatures.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO10 Mean Temperature of the Warmest Quarter
}
