% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio11.R
\name{bio11}
\alias{bio11}
\alias{cpp_bio11}
\title{Mean Temperature of Coldest Quarter}
\usage{
cpp_bio11(tas)

bio11(tas, also.quarter = FALSE)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average temperature of the coldest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Mean Temperature of Coldest Quarter

BIO11 Mean Temperature of Coldest Quarter
}
