% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio12.R
\name{bio12}
\alias{bio12}
\title{BIO12 Total Annual Precipitation}
\usage{
bio12(pr, also.quarter = FALSE)
}
\arguments{
\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO12 Total Annual Precipitation
}
