% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio13.R
\name{bio13}
\alias{bio13}
\title{BIO13 Precipitation of Wettest Month}
\usage{
bio13(pr)
}
\arguments{
\item{pr}{raster stack of monthly precipitation.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO13 Precipitation of Wettest Month
}
