% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio15.R
\name{bio15}
\alias{bio15}
\title{BIO15 Precipitation Seasonality}
\usage{
bio15(pr)
}
\arguments{
\item{pr}{raster stack of monthly precipitation.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO15 Precipitation Seasonality
}
\details{
Seasonality is defined as the coefficient of variation, 
i.e. sd(pr) / divided by mean(pr)
}
