% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio16.R
\name{bio16}
\alias{bio16}
\alias{cpp_bio16}
\title{Mean Precipitation of Wettest Quarter}
\usage{
cpp_bio16(pr)

bio16(pr, also.quarter = FALSE)
}
\arguments{
\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average precipitation of the wettest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Mean Precipitation of Wettest Quarter

BIO16 Mean Precipitation of Wettest Quarter
}
