% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio17.R
\name{bio17}
\alias{bio17}
\alias{cpp_bio17}
\title{Mean Precipitation of Driest Quarter}
\usage{
cpp_bio17(pr)

bio17(pr, also.quarter = FALSE)
}
\arguments{
\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average precipitation of the driest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Mean Precipitation of Driest Quarter

BIO17 Mean Precipitation of Driest Quarter
}
