% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio18.R
\name{bio18}
\alias{bio18}
\alias{cpp_bio18}
\title{Mean Precipitation of Warmest Quarter}
\usage{
cpp_bio18(tas, pr)

bio18(tas, pr, also.quarter = FALSE)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}

\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average temperature of the wettest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Mean Precipitation of Warmest Quarter

BIO18 Mean Precipitation of Warmest Quarter
}
