% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bio19.R
\name{bio19}
\alias{bio19}
\alias{cpp_bio19}
\title{Mean Precipitation of Coldest Quarter}
\usage{
cpp_bio19(tas, pr)

bio19(tas, pr, also.quarter = FALSE)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}

\item{pr}{raster stack of monthly precipitation.}

\item{also.quarter}{TRUE/FALSE (default FALSE), if to return also the 
starting month of the warmest quarter.}
}
\value{
matrix with average temperature of the wettest quarter (first
  column) and the first month of the quarter (second column).

one raster with the bioclimatic variable.
}
\description{
Mean Precipitation of Coldest Quarter

BIO19 Mean Precipitation of Coldest Quarter
}
