% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim.R
\name{bioclim}
\alias{bioclim}
\title{Bioclimatic Variables}
\usage{
bioclim(tas, tmin, tmax, pr, checks = FALSE)
}
\arguments{
\item{tas}{raster stack of monthly average temperatures.}

\item{tmin}{raster stack of monthly minimum temperatures.}

\item{tmax}{raster stack of monthly maximum temperatures.}

\item{pr}{raster stack of monthly precipitation.}

\item{checks}{boolean, to run some checks on input layers.}
}
\value{
a raster stack.
}
\description{
Bioclimatic Variables
}
\details{
This calculates all bioclimatic variables.
}
