% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_layers.R
\name{check_precipitation}
\alias{check_precipitation}
\title{Check precipitation layers}
\usage{
check_precipitation(p)
}
\arguments{
\item{p}{timeseries of precipitations.}
}
\value{
NUll
}
\description{
Check precipitation layers
}
