% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_layers.R
\name{check_temperature}
\alias{check_temperature}
\title{Check temperature layers}
\usage{
check_temperature(t, K0 = -273.15)
}
\arguments{
\item{t}{timeseries of temperatures.}

\item{K0}{absolute zero in Celsius.}
}
\value{
NUll
}
\description{
Check temperature layers
}
