% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{highest_quarter}
\alias{highest_quarter}
\title{Temperature Of The Highest Quarter}
\usage{
highest_quarter(x)
}
\arguments{
\item{x}{matrix of timeseries.}
}
\value{
matrix with average temperature of the highest quarter (first
  column) and the first month of the quarter (second column).
}
\description{
Temperature Of The Highest Quarter
}
