% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lowest_quarter}
\alias{lowest_quarter}
\title{Average Value of the Lowest Quarter}
\usage{
lowest_quarter(x)
}
\arguments{
\item{x}{matrix of timeseries.}
}
\value{
matrix with average temperature of the lowest quarter (first
  column) and the first month of the quarter (second column).
}
\description{
Average Value of the Lowest Quarter
}
