% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio14.R
\name{bio14}
\alias{bio14}
\title{BIO14 Precipitation of Driest Month}
\usage{
bio14(pr)
}
\arguments{
\item{pr}{raster stack of monthly precipitation.}
}
\value{
one raster with the bioclimatic variable.
}
\description{
BIO14 Precipitation of Driest Month
}
